Dockerfile fixed
The Dockerfile creates an image with Datacube and Datacube Server.

It also creates the default folders for .fits and .nc files and downloads the cube 1342228703_M17-2_SPIRE-FTS_15.0.3244_HR_SLW_gridding_cube.fits that is opened at startup.
These folders can be replaced with a host one using the -v option to bind-mount one to /data as written in the "run.sh" script.
Option -Xmx5G pour DataCubeServer dans docker