Localisation to Location
Use "Location" and not "Localisation" that is spelled incorrectly in english ("Localization") .

Location = a particular place or position.
Localization = the process of making something local in character or restricting it to a particular place.
A pull request will follow in DataCubeServer to match this change.

Also changed "type object" to "object type"
and removed unused "NameLocalisation" interationalization key