Spectre and Histogram switch place
When reading the spectre, one would like to see their emplacement on the image easily without having to scroll up and down all the time.
To fix this we put the Histogram down in place of the Spectre.
This fixes the redmine ticket 2889 - https://idoc-projets.ias.u-psud.fr/redmine/issues/2889