Edit navbar style
Removed About as a page and changed the tree bars to a cog. 
There were a few issues with switching from home to about :

-loss of previous settings
-bug MizarWeb/MizarWeb.github.io#33
-app freezes when selecting a new file while on the about page
-enhancement MizarWeb/MizarWeb.github.io#14

Having About as a dropdown already loaded doesn't cost much since it just contains text. It seemed overkill to have to load it as a page.